org: patch to indent tables with org-modern
The patch changes org-indent-set-line-properties to change how the line
and wrap prefix are set. If in a table, change those prefixes to include
a smaller font height. When the height is changed, the width changes too
and we need to recalculate the size of the prefix to match the changed
width.